requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Mail
psycopg2-binary
nltk
tensorflow
tensorflow-hub
PyPDF2
python-docx
requests
Werkzeug
gunicorn